Data pooling for LODs refactored
